+++ src/main/java/nextstep/subway/domain/fare/FarePolicy.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.fare;
 
 public interface FarePolicy {
     int getAdditionalFee();
