+++ src/main/java/nextstep/subway/domain/path/Path.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.path;
 
 import nextstep.subway.domain.entity.Section;
