+++ src/main/java/nextstep/subway/domain/path/PathFinder.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.path;
 
 public interface PathFinder {
 	Path getShortestPath(Long source, Long target);
