+++ src/main/java/nextstep/subway/domain/path/SectionWeightedEdge.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.path;
 
 import nextstep.subway.domain.entity.Section;
 import org.jgrapht.graph.DefaultWeightedEdge;
